Reproduce: @SkipObscure not working on dict.get
Summary:
This diff reproduces the problem in T128482793, by adding a duplicate model for
method `dict.get`. The root cause is that, we have several models for
`dict.get` and the model join is wrong. Specifically, one model contains mode
`Obscure` and the other model removes this mode (due to SkipObscure). Then,
when joining the models, we get mode `Obscure` in the result, but we actually
want a model without `Obscure` due to SkipObscure. Hence, the join of ModeSet
is wrong w.r.t. “Obscure”.

Reviewed By: arthaud

Differential Revision: D38541658

fbshipit-source-id: 461679dcc91c84b0b64642895bb0e481c6caf417