Add a model for dict.get
Summary:
This is the 2nd part of the fix to the problem in T128482793.

Although method `dict.get` is inherited from class `typing.Mapping` and we have
a model for `typing.Mapping.get`, the call to `dict.get` gets resolved
precisely as `dict.get`, not `typing.Mapping.get`. Hence, the model of
`typing.Mapping.get` is not used.

Reviewed By: arthaud

Differential Revision: D38557465

fbshipit-source-id: 38ac35402211e6df0a66dedb517e4ae914d700e8