Move pyre init "sub" methods [entry point separation 1/3]
Summary: The command pysa-init calls `(pyre) init` initialization commands to prevent reduplicating code. However pyre doesn't require taint models directory while pysa still relies on it. As a first step of decoupling both, moving commonly used code to a methods in `initialize.py`. This will help in the next few diff and is a modularisation effort to prevent code duplication in the future.

Reviewed By: arthaud

Differential Revision: D38541008

fbshipit-source-id: 160838b8a1ec14abb92fbcb2826fd64b42106da4