lazy tracker ls-cache: optimize data format
Summary:
In the previous commit, we implemented a naive cache for the
results of reading directories and converting the paths to
ModulePath values; this gave us a 10x speedup in the lazy module
tracker which is good but not sufficient.

In this commit, we follow that change up by putting the data into
a map where the ModulePath values are grouped by qualifier. In doing
so, we trade many list scans for a single slightly-more-expensive
scan that builds up the map.

Reviewed By: grievejia

Differential Revision: D38537905

fbshipit-source-id: 30813f691d331c04089c13fd584150833d81722d