lazy tracker ls-cache: Implement the cache
Summary:
The cache itself is very simple, and most of the other changes
are just wiring a `LazyFinder.t` value through all the code without
any logic change.

The one other critical change is that we add an extra
`process_module_path_updates` callback to `Layouts.Api.t`, which
is a no-op in the eager case but we use to handle invalidation for
lazy layouts; we invalidate a qualifier and all its recursive parents
when we do see an update. This is needed because watchman does not
tell us about directory removals, so the safest way to handle a file
update notification from watchman is to assume it may also correspond
to directory changes at all layers.

**Performance**

It appears that for fat directories we get about a 9-fold speed increase from this change.

Reviewed By: grievejia

Differential Revision: D38529084

fbshipit-source-id: 133947962320343ac9dd112ff00a0c2d1e356fe8