lazy tracker ls-cache: push ModulePath.create down
Summary:
This commit prepares for adding a directory read cache to the lazy
module tracker code by rearranging things so that the operation of
looking up all ModulePaths nested underneath directories corresponding
to a given qualifier is isolated, and dependend upon by both
`find_module_paths` and `find_submodule_paths`.

To do this, I added the `ModulePath.create` step into the existing
`directory_children`, and renamed it to what I consider to be a slighly
better name, `find_inner_module_paths`. It's a little hard to come up
with a name that precisely describes what this function does, so I also
added a function header comment giving a plain-English description.

It's this function that we will cache in our initial optimization!

Reviewed By: alicezhu17

Differential Revision: D38529080

fbshipit-source-id: 4e8729bf104e8394b478026f4115010f4c265de3