lazy tracker ls-cache: add find_submodule_paths
Summary:
We have early performance numbers indicating that Pyre performance
degrades badly with a lazy module tracker in some cases.

This is almost certainly due to "fat" directories with huge numbers
of sibling modules; the current lazy tracker is not optimized against
this, and in particular there are two slightly different problems:
- we have to execute many, many duplicate filesystem reads if a lazy
  load requires loading many of the siblings
- each read is slow, because it returns a lot of data and then we have
  a linear-time search for relevant files

The solution we intend to try is to cache. Partly because we want to
understand the tradeoff I'll do this in two phases:
- first, naively cache directory reads and make sure the cache
  invalidation logic is carefully reviewed and tested. This will tell
  us how much gain we get just by optimizing syscalls
- then post-process the directory reads into a map instead of a list,
  to get better amortized compute costs (the list scans are wasteful);
  this is relatively easy to do once ^ is finished, and it will be
  interesting to measure how much difference it makes.

As is often the case, this change will involve first refactoring existing
logic until we can isolate the appropriate operations, and then implementing
the cache.

This commit kicks off refactoring by moving the logic to find submodule
paths out of the ImplicitModules module and into LazyFinder; this makes
it easier to refactor the internals of LazyFinder.

Reviewed By: alicezhu17

Differential Revision: D38529082

fbshipit-source-id: e80eeb121e434e497eca396ecc7bed81f8a408a0