Minimize key size for lazy Tracker directory-read cache
Summary:
An idea from code review was to reduce the memory footprint of the
lazy tracker's inner module cache by using only the last part of
each reference as the key in the map.

This seems like a good idea, although it costs a little bit of boilerplate
in the code so I broke it out so that we could consider it separately.

Reviewed By: grievejia

Differential Revision: D38606383

fbshipit-source-id: 5ef48657ec1d5b4954011592cf5958a91a0b68e4