Add an API for source and sink filtering in model parsing (2/n source-sink filtering)
Summary:
See the previous diff for context.

In this diff, we add a proper API for source and sink filtering in model parsing, using the new module `SourceSinkFilter`.
This can be created from a configuration, and is passed around (instead of passing `sources_to_keep` and `sinks_to_keep`).
It also provides functions to check if a source or sink should be kept, puting it in a central place.

This diff also moves the set of rules to filter in the configuration. The plan is to have the set of sources and sinks to filter stored there as well.

Finally, this includes a fix for parametric sources/sinks. See the test for detalis.

Differential Revision: D38679721

fbshipit-source-id: aa1cd2dd0ea46fdc864a1ed6f91b9fb5f4ef0bb7