Add options to filter sources and sinks (1/n)
Summary:
# Context
We currently have an option to run pysa with a specific set of rules, using `--rule <CODE>`. This effectively ignores all models with sources and sinks not mentioned in those rules.
The following set of diffs adds options `--source` and `--sink` to run pysa with a specific set of sources and sinks, ignoring all other models.

# This diff
In this diff, we implement the parsing of `--source` and `--sink` to threading the value to the analyzer. The actual handling of the options will be in a follow-up.

Reviewed By: dkgi

Differential Revision: D38679486

fbshipit-source-id: f2969f1ea984a1f5144adead62b2c1a2aea86e9e