Don't use the `Coroutine` wrapper for stubs returning `AsyncIterator`.
Summary:
Problem: When a function is async and not a generator, we turn the return type of the function to be `Coroutine[Any, Any, <original return type>]`. We check if a function is a generator by seeing if its body has `yield`, etc.

https://www.internalfb.com/code/fbsource/[032f32921897]/fbcode/tools/pyre/source/ast/statement.ml?lines=1875-1882

That means: async stubs inadvertently get wrapped in an extra `Coroutine`. Not sure why we haven't seen reports of other issues. Maybe there aren't too many async stubs?

In particular, if there is an `abstractmethod` base stub, the overriding method will get a spurious override-mismatch error.

Reviewed By: grievejia

Differential Revision: D38628247

fbshipit-source-id: accb5a63c283aff86f29ca8f9e457b67d65fe2e2