s/do_not_ignore_errors_in/only_check_paths/g + add flag to oss
Summary:
The name "do no ignore errors in" is pretty confusing, this flag *overrides*
the default source directories we would pass to the backend.

Let's use a better name, and let's include it in the open-source client
as a hidden flag so that if anyone is blocked on slow checks without
watchman we'll have a way to support them.

Reviewed By: grievejia

Differential Revision: D38727135

fbshipit-source-id: 68fb9a0459df25da27523b24bf8b10e1de835ae2