Make a record type more explicit
Summary:
I'm working on extending our tracing dramatically so that our
trace event profiling can provide a few orders of magnitude more
detail about where exactly pyre is slow.

For the moment, I'm not yet sure how to cleanly add the profiling
in a way we would be happy committing to prod, but as a result
it's important that I keep the changes that have to be manually
applied to a minumum.

For whatever reason, when I added profiling to `forward_expression`
ocaml could no longer infer the type of this record without an
explicit module name. So I just want to slip that fix into trunk;
we shouldn't need to mess with irrelevant lines when trying to add
profiling.

Reviewed By: grievejia

Differential Revision: D38747032

fbshipit-source-id: c8790b7b59876eb14cc5202ace41ea534e489cb3