Move the source/sink filtering algorithm in its own module
Summary:
This diff moves the logic that removes sources and sinks that can never match a rule inside its own module.
Since this logic relies on the taint configuration, we stored information for that logic (`matching_sources`, `matching_sinks`, etc) in the taint configuration itself. This is not idea since that information could go out of sync with the configuration itself.
By moving the information is its own module `SourceSinkFilter` and making the type `t` opaque, we force the user to call `SourceSinkFilter.create` with the sound algorithm.
Another benefit is that we could use this to replace the `SourceSinkFilter` from `ModelParser` (see follow-up diff).

Differential Revision: D38765461

fbshipit-source-id: 9f8767afc9516b64998f8322f0ac0fdfdaaf1f15