Fix go-to-def for string annotations.
Summary:
Problem: `x: "Foo"` gets qualified as `x: "test.Foo"`. Then, it gets parsed as `x: test.Foo`.

But, in the second step, we end up (a) shifting the start column of the overall expression by 1, and (b) moving the start of `Foo` by the length of the qualifier.

This causes go-to-def on `Foo` to return results for `test`. SMH.

Solution: After parsing the string annotation, set all locations to be the location of the initial annotation.

This works for us, because when visiting, `test.Foo` is preferred over `test`. So, if their locations are equal, we will get the expected behavior.

Reviewed By: stroxler

Differential Revision: D38788553

fbshipit-source-id: b41fe56d1601695a0aed7dbd89b2b9c287448825