Demonstrate how our handling of string annotations leads to an off-by-one error in locations.
Summary:
Problem: `x: "Foo"` gets qualified as `x: "test.Foo"`. Then, it gets parsed as `x: test.Foo`.

But, in the second step, we end up (a) shifting the start column of the overall expression by 1, and (b) moving the start of `Foo` by the length of the qualifier.

This causes go-to-def on `Foo` to return results for `test`. SMH.

This diff just demonstrates the behavior.

Next diffs: Fixes.

Reviewed By: stroxler

Differential Revision: D38788557

fbshipit-source-id: 29a20510b31534a570c1a1f5c6bf666a73f77833