Fix go-to-def for type annotations.
Summary:
Problem: We treated type annotations as atomic expressions. This meant that, for `Foo[Bar]`, we could neither get the definition for `Foo` nor the definition for `Bar`.

(We didn't get the definition for `Foo[Bar]` because it's not a `Name` - it's an arbitrary expression.)

Visit the annotation expression as an expression, so that we find the narrowest spanning symbol.

Reviewed By: stroxler

Differential Revision: D38788554

fbshipit-source-id: 21391e711fb2b809b73290ac02fa208bab1de72b