Standardize persistent.py names: s/state/server_state/g
Summary:
Inside of the PyreServerHandler, the server state is known by the
variable name server_state, but inside of PyreServer it is known
as `state`.

While I recognize that in some sense `server.server_state` is redundant,
the ability to use a code search to see the connections here is *critical*
for rapidly navigating the codebase; this is a big file that is using
global variables for the most important control flow, and that means that
our naming conventions should be driven by what will make navigation
across that queue-based control flow easiest.

Reviewed By: grievejia

Differential Revision: D38767953

fbshipit-source-id: b1a4f26b4102c55c741020abe2ac1941f2484e45