Use LSP response for hover query
Summary:
No one is using the hover query at this point, so it is safe
to make a breaking change to the format.

The client-side code is significantly easier to write if we don't
*unnecessarily* have the ocaml-side json format diverge from LSP
format; for example, the server go-to-def data is compatible with
the LSP standard and this simplifies the control flow for go-to-def
on the client side, preventing us from needing a whole hierarchy
of otherwise-useless dataclases just to convert formats.

The hover query is not doing this - we currently label the hover
output as `message`, which is inconsistent with the `contents`
key used in LSP.

Let's fix this, it will make life easier at no cost.

Reviewed By: grievejia

Differential Revision: D38772214

fbshipit-source-id: 8222548974fddb2b759dff98de54f8e5318a3a36