Raise a NotImplementedError instance, not the type
Summary:
You can raise anything in python, which is why this isn't a type
error, but we're only really supposed to be raising exceptions.

The code `raise NotImplementedError` is actually raising a type,
not an instance of `Exception`; we want to call the constructor!

I noticed this while I was manually testing the new hover logic;
I hovered on one of these lines and saw `Type[NotImplementedError]`
which made me squint.

Reviewed By: grievejia

Differential Revision: D38791718

fbshipit-source-id: 43c873045f4a8a960e6f4e4dee029c835ce5d948