Hook up HoverQuery to handle hover requests (minimal)
Summary:
In this commit, I flip the handler for hover queries over to using HoverQuery
to get a non-batched server response.

This code leaves quite a bit of dead logic around TypesQuery and still triggers
no-longer-needed queries; I'll clean that up in the next commit, splitting the
cut-over and cleanup makes it easier to both write and review the changes because
quite a lot of test boilerplate needs adjustment once we get rid of the cached
types data.

Reviewed By: grievejia

Differential Revision: D38787259

fbshipit-source-id: 74ae8be4747620b6260acdcf94b44d898e756a1a