Clean up orphaned logic from batched hover
Summary:
This commit gets rid of TypesQuery and the map from path
to types data, cleaning up multiple unnecessary tests along
the way.

This should be a major step forward in making go-to-def fast,
because the types queries are quite expensive compared to a
single hover request on a large file.

Reviewed By: grievejia

Differential Revision: D38787258

fbshipit-source-id: 8296c88b1f84840bd145818fc0a7e3e81e0ee82b