Deobfuscate persistent.py: make it clearer that ServerState is final
Summary:
The naming conventions around `server_state` as well as the comment on it
suggested that the reference itself inside of `PyreServer` might be mutable.

This is not the case - `server_state` is final. Only the fields inside of it
are mutable. Making `PyreServer` a frozen dataclass makes this much clearer
(and enforced by the type system).

Reviewed By: grievejia

Differential Revision: D38768147

fbshipit-source-id: 318e48b74ef21f52584ed5b9149d35e2a4d444ca