Clarify names in persistent: s/state/query_state/g
Summary:
The persistent code is pretty hard to follow right now.

Instead of being handled by normal call stacks, all control flow runs
through various global state variables.

There are good historical reasons for this; we are hoping to reduce the
extent of the indirection. But in the meantime there are *not* good reasons
for the names of those critical global variables to differ depending where
we are in the codebase - global variables need to have the same name everywhere,
or my head will explode.

Here, I rename PyreQueryHandler.state to PyreQueryHandler.query_state;
this is the name that it has in the PyreServer, where `state` refers to
something else. Using the same name everywhere makes navigation across
the global queue quite a bit easier, I can just search for the variable
name to see where data is enqueued vs dequeued.

Reviewed By: grievejia

Differential Revision: D38767284

fbshipit-source-id: c428fd2617d486f75112c4975b7527c951081141