Add optional command to skip environmental setup [entry point migration 4/6]
Summary:
The pysa-init command sets up an environment to run pysa in addition to setting up a .pyre_configuration file. Make that setup optional as interally it doesn't make much sense to do so.

As part of the plan to remove setting up taint_models_directory in pyre init, re-adding pyre init-pysa command to the internal client.

Reviewed By: arthaud

Differential Revision: D38704446

fbshipit-source-id: 407dded1ca6be70679d1bd3cea48e174a28c5afa