Improve source and sink filtering using command line arguments
Summary:
During the analysis, every time we apply tito, we filter sources and sinks that can never match a rule.
This diff improves the algorithm to take into account the `--source`, `--sink` and `--transform` command line arguments.
In a follow-up, we will finally replace `ModelParser.SourceSinkFilter` with `TaintConfiguration.SourceSinkFilter` since the latter one is more powerful.

Reviewed By: alexkassil

Differential Revision: D38809962

fbshipit-source-id: 2359a686801e0669658210885da5a3ac69e055d2