Pyre Correctness Test: Add Core Logic to run_unsaved_changes_test
Summary:
**Context**

This diff is to setup the core logic for `run_unsaved_changes_test`. We want to test unsaved changes against incremental changes and see if they differ on a per-file basis.

**Procedure**

The procedure we use for testing is to grab commits from this directory: https://www.internalfb.com/code/fbsource/fbcode/tools/pyre/source/command/test/integration/fake_repository/

We'll go through each of the commits one by one.
1. Initialize a directory with the same files with empty content.
2. Copy over each file one by one and run an overlayUpdate, get the pyre errors.
3. Send an incremental update (analogous to saving the file), and get the pyre errors.
4. Compare the pyre errors from the overlay & the incremental update.

**Not Done: Pyre Full Check Comparison**

I considered/tried implementing full check errors to compare to overlay updates & figured that would be a nice followup. However, the errors have slightly different formats which makes them difficult to compare. Also, we've basically verified incremental to be the same as full check, so I feel it's safe to put off this enhancement until the future.

**Next Step**

Actually implement the body for the different `Repository` methods.

Reviewed By: grievejia

Differential Revision: D38435960

fbshipit-source-id: 874f2f659c49926b8070be1c98665d8a1bc61d4e