Refactor persistent.py: deduplicate try-connect logic
Summary:
The handling of connections after a successful server start
is exactly like the handling when connecting to an existing
server, except for a single log message.

If we just parametrize the logging, we can deduplicate all
the critical logic which makes it much clearer that the
connection itself is identical.

Reviewed By: grievejia

Differential Revision: D38851943

fbshipit-source-id: abdcd580999905c6e1b7188207cf550ca4c0a7da