refactor persistent.py: simplify control flow
Summary:
In the previous commit, I deduplicated the try-connect logic
for existing versus new servers while preserving the behavoir
*exactly*, and the "I connected to a new server" message was
being logged *before* the connection.

The resulting logging code is messy, it's harder to see what
is really going on when we log the two events in different ways.

This commit proposes to log the success after the connection.

I split it into a separate commit because we might consider it
important the order of logging in case the connection fails
after the server started successfully, and therefore worth the
messier code; this code is easier to read but the quality
of our logs is more important; it boils down to whether we think
the lack of a socket connection is common enough to worry about
(beyond stack traces, which will show us what happened in any
case; one reason I favor the simplification is that we would
probably just rely on the stack trace anyway, and I don't think
the separate "Pyre server has been initialized" message really helps
any)

Reviewed By: grievejia

Differential Revision: D38865351

fbshipit-source-id: 10c5f87083523418b5040de4c8fd3a2aeba898b6