Refactor persistent.py: use consistent logging
Summary:
I am working toward deduplicating logic in PyreServerHandler
so that I can more easily skim the control flow.

One of the problems is that the logging and status showing code
isn't always in the same order, and we don't always log the same
data.

But I don't beleive the order really matters, and I can't see
any good reason for us to *not* log the exception on the final
failure; if anything, I would *most* want the exception on
the last event since that gives us deduplicated exceptions.

So, let's log the same data on SUSPENDED that we log on NOT_CONNECTED,
and let's log before we send the user information.

This will allow me to deduplicate logic in the next commit.

Reviewed By: grievejia

Differential Revision: D38851945

fbshipit-source-id: 560fa40c6dffbc7568197b50311b8cfb26154e4a