Rename connect functions in both connection modules
Summary:
I am about ready to combine the async and sync connection modules
into a single module, where the types and function names can distinguish
sync vs async connections rather than us relying on the module names -
which we routinely mangle in import statement, which makes them harder
to follow - to handle this.

The last step is simply to use different names for the connection
functions.

Reviewed By: grievejia

Differential Revision: D38883871

fbshipit-source-id: 820a67bebfb9840eb7e575ae4f26bbd32ccff6be