Improve naming: split socket connection path logic out
Summary:
The logic in `server_connection` and `async_server_connection`
for synchronous and aynchronous I/O is not pyre-specific. I'd
like to keep the naming clearly separated from the pyre-specific
socket path logic (I'm actually hoping to combine the two connection
modules in an upcoming commit).

To this end, I'm moving the socket functions into a new module
named `daemon` (also at this point I'm moving forward with my plan
to start referring to the ocaml backend server as a "daemon" in
frontend code, so that we can avoid the confusion caused by
overloading of the term "server" in an IDE-first world).

Reviewed By: grievejia

Differential Revision: D38878042

fbshipit-source-id: 7a9fb7e77b4999dab1e20c73589bfcd984d8362e