Make a `connections` module, get rid of module renames
Summary:
We now have a single `connections` module where the type
and function names clearly distinguish async vs sync connection
logic.

This allows me to get rid of the `async_server_connection` to
`connection` renames, which made it harder for me to navigate
the code.

It also gets rid of the `server` in these names - I
found this misleading because it seemed to carry connotations
that these connections are always to the daemon, but in reality
they are sometimes connections to the *editory* (i.e. they
are just async stdin and stdout), which surprised me very
much.

Reviewed By: grievejia

Differential Revision: D38883868

fbshipit-source-id: e01e929feee68bd0526abc7f11448212d4e524ce