Move the shared query logic into `daemon.py`
Summary:
The `daemon` module should know not only how to get a socket path,
but how to use one and execute queries.

This commit moves the synchronous query api logic, as well as the
lower-level serialization code, from `query.py` (which I want to
be focused on how to run `pyre query`) into `daemon.py`.

In the process I also deduplicate some of the serialization code
that had become copypasta, and move it into static methods of the
Response class.

Reviewed By: grievejia

Differential Revision: D38885981

fbshipit-source-id: 332542951268ce945ce6271b64ec1bc00da79bab