Do not analyze test files using pytest
Summary:
We generally exclude test files from the taint analysis.
Test files are currently detected by checking if a class inherits from `unittest.TestCase`. This means we only handle source code using the `unittest` framework.
This diff adds a heuristic that checks for files using `pytest`. If a file imports `pytest` and defines at least one function starting with `test_`, we will consider it a test file and ignore it for the taint analysis.

Differential Revision: D38930712

fbshipit-source-id: 65a054f35d74aa4177fb5da41e3ab2818d8999f7