Extract all of the start options loading into a callback
Summary:
In review of an earlier diff, it was suggested that we can lock down the
behavior of the language server code by not passing *any* arguments directly.

The idea was inspired by thinking that the arguments are only used to load
configuration, which is not true - we need StartArguments to create the backend
server start args as well.

But the idea is nontheless solid: if we extract *all* of the delayed-evaluation,
.pyre_configuration - dependent code into a callback, we don't need to pass
any arguments structs at all.

This commit makes that change. In the next commit I'll explore one of the
implications, which is that we can move the construction of a StartArguments
dataclass from CommandArguments (which is kind of painful due to the 80% duplication
in fields) back down into `persistent.py`.

Reviewed By: grievejia

Differential Revision: D38999740

fbshipit-source-id: f572d5bcd23e7bab0bc596aef775905e9018a503