Lift the configuration loading logic into `pyre.py`
Summary:
In order to allow us to spin up a shadow-mode language server
using alternative daemon startup flags (in particular, laziness),
I want the `persistent` code to be able to accept a `StartArguments`
rather than a `CommandArguments` - at the moment it is hardcoded
that we have a nonlazy server with no type check resolution, etc.

Before I can do that, I have to deal with another issue: the construction
of a `configuration_module.Configuration` accepts a `CommandArguments`
as input. For most commands, we just build the configuration up front and
use it, but in language servers we want the ability to reload the config
from disk. As a result, we need to be passing down a callback to do this.

A bonus is that now we no longer have a mysterious `base_directory` argument
floating around the `persistent` codebase - it's more immediately obvious
that the only thing we were using it for was reloading config.

Reviewed By: grievejia

Differential Revision: D38994075

fbshipit-source-id: 0d7cbab17b0e071596a8fbced5abce8989618bf0