Use StartArguments in persistent.py
Summary:
In order to create a code-nav server that exposes additional
flags (for example, use lazy module tracking) and may hardcode other
behaviors that differ from the legacy pyre language server (for example,
cache type annotations in the TypeEnvironment), we want to be able
to control the arguments sent to the daemon's start command.

In order to do this, `persistent.py` needs to accept a `StartArguments`
as input rather than constructing it from CommandArguments (with hardcoded
flag values that are wrong for the code navigation use case).

This commit makes that change.

Reviewed By: grievejia

Differential Revision: D38994076

fbshipit-source-id: f7ca11f8a42f0828adfaab3a32477337efa5659f