Move PyreServer down
Summary:
We generally like our python code to be ocaml-style, with
low-level logic at the top and high-level logic at the bottom.

I don't have strong opinions about whether this is the best way,
but I do think being consistent helps with code navigation: the top-level
entity inside a persistent instance is the PyreServer (we've talked about
maybe splitting it into peices, but even if we do this logic is clearly
higher-level than the handler logic is).

Let's move it to the bottom, it will mean less jumping around when trying
to understand the server launch control flow in particular.

Reviewed By: grievejia

Differential Revision: D39023280

fbshipit-source-id: 382ad86d183ea88531e15002633f95d282c848d3