Rename the `_run` methods
Summary:
I want more specific names for the looping functions than `_run`,
which is particularly irritating to me given that we have a `run`
method on each of the classes with a `_run` method and they don't
do the same things.

The outer `run` methods probably shouldn't be renamed, since they expose the
top-level functionality and it helps with interfacing with the rest of the code
to have a standardized name for the top-level. But the lower level function
can have a descriptive name; I do that for each of the three runnable classes
in this commit.

Reviewed By: grievejia

Differential Revision: D39025277

fbshipit-source-id: 1b814790559bbbfd1b30a494b7477bb3ed5321e8