Make a smart constructor for StartArguments
Summary:
We can cut down on quite a bit of boilerplate, and keep the argument
construction logic in top-level click entrypoints (which I think is
mostly a good idea - the whole point of these dataclasses is to zip
up various click-informed arg info) by using a smart constructor
so that when we make a `StartArguments`, the fields coming straight
from `CommandArguments` are automatically propagated.

The resulting code is less repetitive and also makes it easier to
see which arguments are *specific* to starting a server.

This commit makes the smart constructor and uses it in the entrypoint
for `persistent`, to show how it works.

I'll flip all the other entrypoints in a separate diff; I want this
one to stay small because that makes it easier to see the change and
the effect.

Reviewed By: grievejia

Differential Revision: D39002774

fbshipit-source-id: 676f60168f37cbfa397506957458d4ec8abd50ba