Rename backend things to use daemon
Summary:
This commit continues my push to disambiguate - at least within
the language server code - things that are about the "language server"
versus the "ocaml server" by refering to the ocaml server as the
"daemon".

In addition, the old `PyreServerHandler` seemed to have a not very descriptive
name; I've now renamed it as the `PyreDaemonLaunchAndSubscribeHandler` which
makes it much clearer what it does (and how that differs from the
`PyreQueryHandler`).

Reviewed By: grievejia

Differential Revision: D39023749

fbshipit-source-id: 62e2b78e460d83b5778f06cb4abb4c19ac9f2f9c