Extract a read_frontend_configuration callback
Summary:
In the interests of allowing open source and the internal client
to eventually produce different `frontend_configuration.Base` types
(which isn't needed today, but might be someday) we want to avoid
hardcoding `frontend_configuration.OpenSource(...)` in the
`persistent.py` code.

To that end, we extract a `read_frontend_configuration` callback which
is in turn passed to the construction of the `read_server_start_options`
callback.

I split this into a separate commit because it does make the commandline
entrypoints more complex, but it also pays down some tech debt.

Reviewed By: grievejia

Differential Revision: D39001858

fbshipit-source-id: 72b1334e14b768d43b6f4caf907f9f5e25cfbae5