Move `get_project_identifier` into `frontend_configuration`
Summary:
This is based on a suggestion in diff review that we should deduplicate
the logic in `start.get_server_identifier` with `daemon_socket.get_project_identifier`.

This does change behavior - the old server identifier used just the base
name of the global root and used a single slash - but only in some logging
messages (not scuba logs, just stderr logs) so it should not be a problem.

The low-level functionality is available as a module-level function as well
as via a method on `Base`.

The function is needed at least temporarily because we create the project
identifier directly from the global + relative local roots in some places, so a
method on `Base` is not sufficient without further refactors. I'll try to clean
that up soon but it might be too hard to do right now.

Reviewed By: grievejia

Differential Revision: D39064457

fbshipit-source-id: e91cc3883848a99c8e0e165d6bd7902759fd089a