Make a single _get_server_status function
Summary:
Instead of returning an optional running server status
and then switching on whether it is None in the upstream code to
create the defunct status, let's just have the function create one
or the other and switch on the type of the result.

This makes the variable names simpler and the types tell us more
about what the actual control flow will be.

Reviewed By: grievejia

Differential Revision: D39044444

fbshipit-source-id: ca5775d3ca96bcd83d12b7adda1f1d3e69bec79e