Add support for looking up attributes from a `ReadOnly` type.
Summary:
To resolve attributes, we use `resolve_class` on the type.

We simply ignore the `ReadOnly` wrapper (just like `Annotated`) in the type checking analysis.

Note: I also added a unit test for `Annotated`, since it was missing. Lmk if anyone has strong feelings about mixing it in this diff.

Reviewed By: stroxler

Differential Revision: D39076861

fbshipit-source-id: 640c030a77c6481ec2036d779046d41b5757ee81