Use only initial sever options in PyreQueryHandler
Summary:
It has never been the case that PyreQueryHandler will magically know to
restart if the configuration changes; PyreDaemonLaunchAndSubscribeHandler
can do this because the daemon will tell it if a critical file changes
via a subscription, but PyreQueryHandler cannot.

This is okay!

The pyre query handler only uses start options to get a few things:

- the socket path: this is fine, it never actually changes (even though
  we get it via the configuration... see my commentary on earlier diffs
  about how this isn't pretty but there may not be a better way)
- some ide_features; I think we can live with these not auto-reloading;
  if someone really wants to toggle them off and on they'll just have to
  reload the vscode window

But using a reader and pretending that the QueryHandler can reload when
it can't really is just adding confusion. More importantly, it makes it
look as though the fact that the QueryHandler is in a background task
really matters but this is not true!

This is important since we want to get rid of the PyreQueryHandler
entirely and just deal with requests using normal async logic in the
PyreServer handler.

-----

Some notes about why putting the handlers directly in the PyreServer,
as this commit is prepping us to do, cannot make things works (in particular I
want to make it clear that PyreQueryHandler did not work the way we were
probably imagining it did, and hence getting rid of it can't really hurt)

- It is not the case today that pyre query handler, on a connection error,
  can restart the server. There is no way for it to talk to
  PyreDaemonLaunchAndSubscribeHandler. If we did want this ability,
  it would be better to have the PyreServer directly connecting
  (as we intend) because it has better ability to restart that
  background task.
- It is not the case today that under normal circumstances a config
  change would lead to the PyreQueryHandler seeing the new start
  options. It doesn't hold open a long-running connection to the server,
  so the server restarting won't generally cause any error and therefore
  it won't generally be restarted. If we do want this ability, we really
  need to be getting it on-the-fly via the server state and relying on
  the PyreDaemonLaunchAndSubscribeHandler to reset the server state
  as needed. Again, this is at least as easy to do via the PyreServer
  than with an unnecessary background task.

Reviewed By: grievejia

Differential Revision: D39079392

fbshipit-source-id: 8e502f47852fa5de925983e6a93f8c73d9f39724