Use consistent and public names for request handlers in PyreQueryHandler
Summary:
I am almost ready to get rid of the background task for PyreQueryHandler, which will
turn it into a normal object that PyreServer uses to handle sending requests to the
daemon.

Before I do that, it makes sense to use more predicatable, and non-underscore-prefix,
names for the various handler methods.

Reviewed By: grievejia

Differential Revision: D39079400

fbshipit-source-id: c3e7fd8776990b1f528d6840a6768338c18cc985