Fix lanaguage server protocol names (to the extent we can)
Summary:
It appears to me we had some mindless copy-paste when we defined
code navigation classes - we kept the "TextDocument" in all of the
request names, but they aren't there in the names of the actual
LSP requests - only "text document" operations like save, change, and
close have that in the name.

https://microsoft.github.io/language-server-protocol/specifications/lsp/3.17/specification/#textDocument_typeDefinition

-----------

It would be awesome if we used the language server protocol specification
to define our lsp class names.

This is not possible if we rigidly adhere to no-abbreviations, which sucks
because it means nothing in our codebase can be googled (the language
server protocol uses "Params", so if you search our class names you won't
find anything).

Reviewed By: grievejia

Differential Revision: D39102217

fbshipit-source-id: 5cfaf2bc4c816f827626efcbe0143772f63753c4