Convert PyreQueryHandler to RequestHandler
Summary:
This commit is the big switch I've been working toward for a while
now: we get rid of the background task that pops requests off a queue
and instead just directly call the various request-handling functions
in straight-line async code.

This means the language server can now open multiple connections to the
daemon, but we believe this is not a problem. And it will make the logic
we need for shadow mode *vastly* easier.

We can also think about deleting RequestHandler entirely. It wasn't possible
to do that in one shot without an enormous diff that would be too hard to
review, but now that we no longer need a background task we can look at moving
the handler functions into static scope and having the PyreServer simply
track the various flags that control behavior.

That work is important but can be deferred; this is the key change that lets
us move forward with shadow-mode behaviors, so I'll focus on that before
doing more "code cleanliness" refactors.

Reviewed By: grievejia

Differential Revision: D39079401

fbshipit-source-id: 1924df24d835f5737ebfcf767f6254de81f54cc6